refactors scheduler code
